first stage of removing NDM from Nethermind solution

+++ src/Nethermind/Nethermind.DataMarketplace.Core/Services/INdmBlockchainBridge.cs
@@ -31,7 +31,7 @@ namespace Nethermind.DataMarketplace.Core.Services
         Task<Block?> GetLatestBlockAsync();
         Task<UInt256> GetNonceAsync(Address address);
         Task<NdmTransaction?> GetTransactionAsync(Keccak transactionHash);
-        Task<int> GetNetworkIdAsync();
+        Task<long> GetNetworkIdAsync();
         Task<byte[]> CallAsync(Transaction transaction);
         Task<byte[]> CallAsync(Transaction transaction, long blockNumber);
         ValueTask<Keccak?> SendOwnTransactionAsync(Transaction transaction);
